poem_agent/pyproject.toml
[project]
name = "poem-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
]
